Crate driver_block

Common traits and types for block storage device drivers (i.e. disk).

Re-exports

pub use driver_common::BaseDriverOps;

pub use driver_common::DevError;

pub use driver_common::DevResult;

pub use driver_common::DeviceType;

Modules

bcm2835sdhcibcm2835-sdhci

SD card driver for raspi4

ramdiskramdisk

Mock block devices that store data in RAM.

Traits

BlockDriverOps

Operations that require a block storage device driver to implement.